+++ MCFCRT/src/ext/wcpcpy.c
@@ -17,7 +17,7 @@ wchar_t *_MCFCRT_wcpcpy(wchar_t *restrict s1, const wchar_t *restrict s2){
 	const wchar_t *erp = s2;
 	__m128i xz[1];
 	__MCFCRT_xmmsetz(xz);
-	int shift = (int)((const wchar_t *)s2 - rp);
+	unsigned shift = (unsigned)((const wchar_t *)s2 - rp);
 	uint32_t skip = (uint32_t)-1 << shift;
 	for(;;){
 		__m128i xw[4];
@@ -25,7 +25,7 @@ wchar_t *_MCFCRT_wcpcpy(wchar_t *restrict s1, const wchar_t *restrict s2){
 		__MCFCRT_xmmload_4(xw, rp, _mm_load_si128);
 		mask = __MCFCRT_xmmcmp_41w(xw, xz, _mm_cmpeq_epi16) & skip;
 		if(_MCFCRT_EXPECT_NOT(mask != 0)){
-			shift = __builtin_ctzl(mask);
+			shift = (unsigned)__builtin_ctzl(mask);
 			ewp = _MCFCRT_rep_movsw(ewp, erp, (size_t)(rp + shift - erp));
 			*ewp = 0;
 			return ewp;
