+++ MCFCRT/src/stdc/wchar/wmemchr.c
@@ -16,7 +16,7 @@ wchar_t *wmemchr(const wchar_t *s, wchar_t c, size_t n){
 	rp = (const wchar_t *)((uintptr_t)rp & (uintptr_t)-64);
 	__m128i xc[1];
 	__MCFCRT_xmmsetw(xc, (uint16_t)c);
-	int shift = (int)((const wchar_t *)s - rp);
+	unsigned shift = (unsigned)((const wchar_t *)s - rp);
 	uint32_t skip = (uint32_t)-1 << shift;
 	for(;;){
 		ptrdiff_t dist = rp - ((const wchar_t *)s + n);
@@ -31,7 +31,7 @@ wchar_t *wmemchr(const wchar_t *s, wchar_t c, size_t n){
 		__MCFCRT_xmmload_4(xw, rp, _mm_load_si128);
 		mask = __MCFCRT_xmmcmp_41w(xw, xc, _mm_cmpeq_epi16) & skip & zskip;
 		if(_MCFCRT_EXPECT_NOT(mask != 0)){
-			shift = __builtin_ctzl(mask);
+			shift = (unsigned)__builtin_ctzl(mask);
 			return (wchar_t *)rp + shift;
 		}
 		rp += 32;
